requirements.txt
beautifulsoup4==4.12.0
requests==2.31.0
PyYAML==6.0.1
fastapi==0.109.0
uvicorn==0.27.0
apscheduler==3.10.1
